Updating to bevy 0.3
