+++ src/lib.rs
@@ -15,7 +15,7 @@ pub struct TiledMapPlugin;
 impl Plugin for TiledMapPlugin {
     fn build(&self, app: &mut AppBuilder) {
         app.add_asset::<map::Map>()
-            .add_asset_loader::<map::Map, loader::TiledMapLoader>()
+            .init_asset_loader::<loader::TiledMapLoader>()
             .add_system(process_loaded_tile_maps.system());
 
         let resources = app.resources();
